refactor: update user adapter folder structure
