+++ src/adapters/controllers/errors/index.ts
@@ -1,0 +1,2 @@
+export * from '@/adapters/controllers/errors/missing-params-error';
+export * from '@/adapters/controllers/errors/invalid-params-error';
