+++ src/adapters/controllers/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from '@/adapters/controllers/interfaces/controller-interface';
+export * from '@/adapters/controllers/interfaces/validation-interface';
