+++ src/adapters/controllers/util/validations/index.ts
@@ -1,0 +1,3 @@
+export * from '@/adapters/controllers/util/validations/validation-composite';
+export * from '@/adapters/controllers/util/validations/required-fields-validation';
+export * from '@/adapters/controllers/util/validations/compare-fields-validation';
