+++ src/adapters/controllers/util/validations/validation-composite.ts
@@ -1,4 +1,4 @@
-import { IValidation } from '@/adapters/interfaces';
+import { IValidation } from '@/adapters/controllers/interfaces';
 
 export class ValidationComposite implements IValidation {
   constructor(private readonly validations: IValidation[]) { }
