+++ src/adapters/interfaces/index.ts
@@ -1,4 +1,1 @@
-export * from '@/adapters/interfaces/controller-interface';
-export * from '@/adapters/interfaces/middleware-interface';
 export * from '@/adapters/interfaces/http-interface';
-export * from '@/adapters/interfaces/validation-interface';
