+++ src/adapters/middleware/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from '@/adapters/middleware/interfaces/middleware-interface';
