+++ src/adapters/middleware/interfaces/middleware-interface.ts
